added mapqMin and mapqMax arguments to qCount/qProfile/qMeth and their various C backends; updated documentation and added unitTest for qCount; eliminated many compiler warnings concerning implicit type conversion
git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/QuasR@80250 bc3139a8-67e5-0310-9ffc-ced21a209358

+++ src/cat_bam.c
@@ -113,7 +113,7 @@ int _bam_cat(int nfn, char * const *fn, const bam_header_t *h, const char* outba
         j=0;
         fp_file=fp->fp;
 #ifdef _USE_KNETFILE
-        while ((len = knet_read(in->fp, buf, BUF_SIZE)) > 0) {
+        while ((len = (int)knet_read(in->fp, buf, BUF_SIZE)) > 0) {
 #else  
         while (!feof(in->file) && (len = fread(buf, 1, BUF_SIZE, in->file)) > 0) {
 #endif
@@ -123,14 +123,14 @@ int _bam_cat(int nfn, char * const *fn, const bam_header_t *h, const char* outba
                     Rf_error("[%s] ERROR: truncated file?: '%s'.\n", __func__, fn[i]);
                     return -1;
                 }
-                fwrite(ebuf, 1, len, fp_file);
+                fwrite(ebuf, 1, (size_t)len, fp_file);
                 memcpy(ebuf,ebuf+len,diff);
                 memcpy(ebuf+diff,buf,len);
             } else {
                 if(j!=0) fwrite(ebuf, 1, es, fp_file);
                 len-= es;
                 memcpy(ebuf,buf+len,es);
-                fwrite(buf, 1, len, fp_file);
+                fwrite(buf, 1, (size_t)len, fp_file);
             }
             j=1;
         }
